chore: create swan package
