+++ packages/accordion/package.json
@@ -48,14 +48,6 @@
         "./sp-accordion-item.js": {
             "development": "./sp-accordion-item.dev.js",
             "default": "./sp-accordion-item.js"
-        },
-        "./src/AccordionBase.js": {
-            "development": "./src/AccordionBase.dev.js",
-            "default": "./src/AccordionBase.js"
-        },
-        "./src/AccordionItemBase.js": {
-            "development": "./src/AccordionItemBase.dev.js",
-            "default": "./src/AccordionItemBase.js"
         }
     },
     "scripts": {
@@ -80,7 +72,8 @@
         "@spectrum-web-components/icon": "^1.2.0",
         "@spectrum-web-components/icons-ui": "^1.2.0",
         "@spectrum-web-components/reactive-controllers": "^1.2.0",
-        "@spectrum-web-components/shared": "^1.2.0"
+        "@spectrum-web-components/shared": "^1.2.0",
+        "@spectrum-web-components/swan": "workspace:*"
     },
     "devDependencies": {
         "@spectrum-css/accordion": "6.0.0-s2-foundations.16"
