+++ packages/progress-bar/tsconfig.json
@@ -6,5 +6,5 @@
     },
     "include": ["*.ts", "src/*.ts"],
     "exclude": ["test/*.ts", "stories/*.ts"],
-    "references": [{ "path": "../../tools/base" }]
+    "references": [{ "path": "../../tools/base" }, { "path": "../../swan" }]
 }
