+++ packages/slider/sp-slider-handle.ts
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { SliderHandle } from './src/SliderHandle.js';
+import { SliderHandle } from '@spectrum-web-components/swan/slider/SliderHandle.js';
 import { defineElement } from '@spectrum-web-components/base/src/define-element.js';
 
 defineElement('sp-slider-handle', SliderHandle);
