+++ packages/slider/tsconfig.json
@@ -6,5 +6,9 @@
     },
     "include": ["*.ts", "src/*.ts"],
     "exclude": ["test/*.ts", "stories/*.ts"],
-    "references": [{ "path": "../../tools/shared" }, { "path": "../number-field" }]
+    "references": [
+        { "path": "../../tools/shared" },
+        { "path": "../number-field" },
+        { "path": "../../swan" }
+    ]
 }
